requirements.txt
# ————————————————————————————————————————————————————————————
# requirements.txt
# ————————————————————————————————————————————————————————————
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
Flask-Migrate==4.0.7
python-dotenv==1.0.1
clerk-backend-api==1.2.0